Fixed default version misleading `Built` reporting
The `built` time is as misleading description because one could thing it's the time at which the build occurs. This is not true, this is really the date at which the commit was done.

To avoid any confusion, replace `Built ...` by `Commit Date ...` which is unambiguous, at the expense of being a bit longer.